requirements.txt
pandas==2.3.3
ipykernel==7.1.0
requests==2.32.5
python-dotenv==1.2.1
howlongtobeatpy==1.0.19
rawg==1.2.0